removed mlx from libft + updated atoi, atol and split + updated word count prototype in ft_count.h

+++ lib_src/lib_count/ft_count_word.c
@@ -6,16 +6,16 @@
 /*   By: ljohnson <ljohnson@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/18 13:35:01 by ljohnson          #+#    #+#             */
-/*   Updated: 2022/01/31 18:03:51 by ljohnson         ###   ########lyon.fr   */
+/*   Updated: 2022/02/09 09:51:16 by ljohnson         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_count_word(const char *str, const char c)
+size_t	ft_count_word(const char *str, const char c)
 {
-	int		a;
-	int		count;
+	size_t	a;
+	size_t	count;
 
 	a = 0;
 	count = 0;
